fix(signup): preserve safe payment recovery and cancellation
