+++ apps/web/app/(auth)/signup/pending-payment/__tests__/provider-runtime-boundary.contract.tsx
@@ -77,6 +77,45 @@ import { createProviderRecoveryFixture } from '@billing-recovery-fixture'
 for (const provider of ['stripe', 'btcpay'] as const) describe(provider + ' runtime-to-rendered recovery', () => {
   afterEach(() => { vi.useRealTimers(); vi.restoreAllMocks(); vi.unstubAllGlobals() })
 
+  it('continues the same payment through the real recovery route after the bounded polling timeout', async () => {
+    vi.clearAllMocks()
+    sessionStorage.clear()
+    authState.pendingSignup = anonymousRecoverySignup({ paymentMethod: provider })
+    vi.stubGlobal('Uint8Array', Object.getPrototypeOf(Buffer.prototype).constructor)
+    vi.stubGlobal('crypto', webcrypto)
+    const fixture = await createProviderRecoveryFixture(provider)
+    const authorityId = fixture.repository.latestSignupPaymentSession()!.id
+    const pending: Promise<unknown>[] = []
+    vi.stubGlobal('fetch', vi.fn((url: string, init: RequestInit) => {
+      expect(new URL(url).pathname).toBe('/auth/signup/payment-session/v2/current')
+      expect(JSON.parse(String(init.body))).toEqual(fixture.request)
+      const request = fixture.app.inject({ method: 'POST', url: new URL(url).pathname, payload: JSON.parse(String(init.body)) })
+        .then(reply => new Response(reply.body, { status: reply.statusCode, headers: { 'content-type': 'application/json' } }))
+      pending.push(request)
+      return request
+    }))
+    const settle = async () => { await act(async () => { await Promise.all(pending) }) }
+    vi.useFakeTimers({ toFake: ['Date', 'setTimeout', 'clearTimeout'] })
+    const view = render(<PendingPaymentPage />)
+    try {
+      await settle()
+      for (let attempt = 0; attempt < 20; attempt++) {
+        await act(async () => { await vi.advanceTimersByTimeAsync(4 * 60_000) })
+        await settle()
+      }
+      expect(fetch).toHaveBeenCalledTimes(20)
+      expect(screen.getByRole('button', { name: 'Check again' })).toBeEnabled()
+      await fixture.repository.confirmPaymentSession(authorityId)
+      fireEvent.click(screen.getByRole('button', { name: 'Check again' }))
+      await settle()
+      expect(screen.getByTestId('step-create-paid-account')).toBeInTheDocument()
+      expect(fetch).toHaveBeenCalledTimes(21)
+      expect(fixture.repository.latestSignupPaymentSession()!.id).toBe(authorityId)
+      expect(fixture.calls().cancel).toBe(0)
+      expect(authState.clearPendingSignupPaymentRecovery).not.toHaveBeenCalled()
+    } finally { view.unmount(); vi.useRealTimers(); await fixture.app.close() }
+  })
+
   it.each(['missing', 'confirmed', 'proof-invalid'] as const)('keeps entered controls through a provider exception, then removes them for %s', async terminal => {
     vi.clearAllMocks()
     sessionStorage.clear()
