+++ apps/web/app/(auth)/signup/use-signup-navigation.ts
@@ -6,6 +6,19 @@ type Checkpoint = { step: string; view: string }
 type Recovery = 'review' | 'setup' | 'payment'
 const KEY = 'silentsuiteSignup'
 
+/** Retire presentation checkpoints only after an exact payment release receipt. */
+export function retireSignupPaymentCheckpoints() {
+  const saved = window.history.state?.[KEY]
+  if (saved?.journey) {
+    try { sessionStorage.setItem(`${KEY}:${saved.journey}`, 'retired') } catch { /* Current history is still retired below. */ }
+  }
+  const state = { ...window.history.state }
+  delete state[KEY]
+  const url = new URL(window.location.href)
+  url.searchParams.delete('recovery')
+  window.history.replaceState(state, '', url.pathname + url.search + url.hash)
+}
+
 /** History contains presentation state only, never proof, password or payment capabilities. */
 export function useSignupNavigation({ enabled, warnOnLeave = true, step, view, restore, intercept }: {
   enabled: boolean
@@ -30,6 +43,7 @@ export function useSignupNavigation({ enabled, warnOnLeave = true, step, view, r
     if (!params.has('token') && !params.has('email_verification_token') && saved?.version === 1) {
       let mutation: string | null = null
       try { mutation = sessionStorage.getItem(`${KEY}:${saved.journey}`) } catch { /* Fall back to the current checkpoint. */ }
+      if (mutation === 'retired') return
       const previousPhase = mutation ?? saved.phase
       setRecovery(previousPhase === 'setup' || previousPhase === 'payment' ? previousPhase : 'review')
     }
@@ -84,9 +98,10 @@ export function useSignupNavigation({ enabled, warnOnLeave = true, step, view, r
       version: 1, journey: journey.current, step, view, phase: next,
     } }, '')
   }
-  const clear = () => {
+  const clear = ({ retirePayment = false }: { retirePayment?: boolean } = {}) => {
+    if (retirePayment) retireSignupPaymentCheckpoints()
     const state = { ...window.history.state }
-    try { sessionStorage.removeItem(`${KEY}:${journey.current ?? state[KEY]?.journey}`) } catch { /* No authority is stored here. */ }
+    try { if (!retirePayment) sessionStorage.removeItem(`${KEY}:${journey.current ?? state[KEY]?.journey}`) } catch { /* No authority is stored here. */ }
     delete state[KEY]
     window.history.replaceState(state, '')
     journey.current = null
